Add CreateIndex function
